--- clog-domain/src/main/kotlin/org/depromeet/clog/server/domain/auth/application/dto/KakaoLoginRequest.kt
@@ -1,5 +1,0 @@
-package org.depromeet.clog.server.domain.auth.application.dto
-
-data class KakaoLoginRequest(
-    val idToken: String
-)
