--- clog-domain/src/main/kotlin/org/depromeet/clog/server/domain/auth/application/dto/LocalLoginRequest.kt
@@ -1,5 +1,0 @@
-package org.depromeet.clog.server.domain.auth.application.dto
-
-data class LocalLoginRequest(
-    val loginId: String
-)
