+++ clog-domain/src/main/kotlin/org/depromeet/clog/server/domain/auth/application/dto/request/LocalLoginRequest.kt
@@ -1,0 +1,5 @@
+package org.depromeet.clog.server.domain.auth.application.dto.request
+
+data class LocalLoginRequest(
+    val loginId: String
+)
