+++ clog-domain/src/main/kotlin/org/depromeet/clog/server/domain/auth/application/strategy/AuthProviderHandler.kt
@@ -1,6 +1,6 @@
 package org.depromeet.clog.server.domain.auth.application.strategy
 
-import org.depromeet.clog.server.domain.auth.application.dto.AuthResponseDto
+import org.depromeet.clog.server.domain.auth.application.dto.response.AuthResponseDto
 
 interface AuthProviderHandler<T> {
     fun login(request: T): AuthResponseDto
